requirements.txt
librosa~=0.10.0
numpy~=1.24.3
pandas~=1.5.2
soundfile==0.9.0
wave~=0.0.2
scikit-learn==0.24.2
tqdm~=4.65.0
matplotlib==2.2.3
pyaudio~=0.2.11
tensorflow~=2.11.0
python-osc~=1.8.1
django~=4.1
keras~=2.10.0
djangorestframework~=3.14.0
channels~=4.0.0
spoutgl~=0.0.5
pygame~=2.5.1
websockets~=10.4